tests: adiciona testes para commands delete, login, update dos commands, produtos, pedidos e cliente

+++ GerenciadorDePedidos.Application/Commands/DeleteCliente/DeleteClienteCommandHandler.cs
@@ -15,6 +15,9 @@ public class DeleteClienteCommandHandler : IRequestHandler<DeleteClienteCommand,
 	public async Task<Unit> Handle(DeleteClienteCommand request, CancellationToken cancellationToken)
 	{
 		var cliente = await _clienteRepository.GetByIdAsync(request.Id);
+		if (cliente is null)
+			throw new KeyNotFoundException("O cliente nao foi encontrado");
+
 		await _clienteRepository.DeleteAsync(cliente.Id);
 		return Unit.Value;
 	}
