+++ GerenciadorDePedidos.Application/Commands/DeleteProduto/DeleteProdutoCommandHandler.cs
@@ -15,6 +15,9 @@ public class DeleteProdutoCommandHandler : IRequestHandler<DeleteProdutoCommand,
 	public async Task<Unit> Handle(DeleteProdutoCommand request, CancellationToken cancellationToken)
 	{
 		var produto = await _produtoRepository.GetByIdAsync(request.Id);
+		if (produto is null)
+			throw new KeyNotFoundException("Produto nao encontrado");
+		
 		await _produtoRepository.DeleteAsync(produto.Id);
 		return Unit.Value;
 	}
